Add files via upload
Association Rules show the relationship between different items in the cart and help us to improve the business buy applying some discount on one item and using that item to increase the sale of other items. This method is also used in Bigmart, Big Bazaar ,etc. to increase their sales figure.